requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.4
langchain>=0.2
langchain-community>=0.2
langchain-text-splitters>=0.2
openai>=1.30
chromadb>=0.5
sentence-transformers>=2.7
pypdf>=4.2
python-dotenv>=1.0
faiss-cpu>=1.8
docx2txt>=0.8
requests>=2.32
